fix parser

+++ src/Commands/slp.js
@@ -8,20 +8,20 @@ module.exports = new Command({
 	name: "slp",
 	description: "Smooth Love Potion",
 	async run(message, args, client) {
-		if (isNaN(parseInt(args[1]))) {
+		if (isNaN(parseFloat(args[1]))) {
 			return message.channel.send("Invalid amount. Type `!slp <amount>` or `!slp <amount> @ <price>`.")
 
 		} else if (args[2] === "@" && !args[3]) {
 			return message.channel.send("Invalid command. Type `!slp <amount> <@> <price>`.")
 
-		} else if (args[2] === "@" && !isNaN(parseInt(args[3]))){
+		} else if (args[2] === "@" && !isNaN(parseFloat(args[3]))){
 			
 			let y = parseFloat(args[3]).toFixed(2);
 			let x = (y * args[1]).toFixed(2);
 
 			message.reply(`\`SLP\` **${args[1]}** => \`PHP\` **${x}**`);
 			
-		} else if (!isNaN(parseInt(args[1])) && !args[2]) {
+		} else if (!isNaN(parseFloat(args[1])) && !args[2]) {
 			
 			fetch('https://api.coingecko.com/api/v3/simple/price?ids=smooth-love-potion&vs_currencies=php')
 			.then(res => res.json())
